[MOVE/#109] 폴더링 수정
