+++ presentation/src/main/java/com/going/presentation/dashboard/DashBoardActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.viewModels
 import com.going.presentation.R
 import com.going.presentation.databinding.ActivityTripDashBoardBinding
 import com.going.presentation.setting.SettingActivity
-import com.going.presentation.entertrip.starttrip.invitetrip.CreateTripActivity
+import com.going.presentation.entertrip.invitetrip.invitecode.CreateTripActivity
 import com.going.ui.base.BaseActivity
 import com.going.ui.extension.setOnSingleClickListener
 import com.google.android.material.tabs.TabLayoutMediator
