+++ presentation/src/main/java/com/going/presentation/entertrip/createtrip/finish/FinishTripActivity.kt
@@ -1,4 +1,4 @@
-package com.going.presentation.entertrip.starttrip.createtrip
+package com.going.presentation.entertrip.createtrip.finish
 
 import android.content.ClipData
 import android.content.ClipboardManager
@@ -8,13 +8,13 @@ import android.os.Bundle
 import com.going.presentation.R
 import com.going.presentation.dashboard.DashBoardActivity
 import com.going.presentation.databinding.ActivityFinishTripBinding
-import com.going.presentation.entertrip.starttrip.createtrip.EnterTripActivity.Companion.DAY
-import com.going.presentation.entertrip.starttrip.createtrip.EnterTripActivity.Companion.END
-import com.going.presentation.entertrip.starttrip.createtrip.EnterTripActivity.Companion.START
-import com.going.presentation.entertrip.starttrip.createtrip.EnterTripActivity.Companion.TITLE
-import com.going.presentation.entertrip.starttrip.invitetrip.InviteFinishActivity.Companion.DATE_FORMAT
-import com.going.presentation.entertrip.starttrip.invitetrip.InviteFinishActivity.Companion.D_DAY_FORMAT
-import com.going.presentation.entertrip.starttrip.invitetrip.InviteFinishActivity.Companion.TRIP_FORMAT
+import com.going.presentation.entertrip.createtrip.choosedate.EnterTripActivity.Companion.DAY
+import com.going.presentation.entertrip.createtrip.choosedate.EnterTripActivity.Companion.END
+import com.going.presentation.entertrip.createtrip.choosedate.EnterTripActivity.Companion.START
+import com.going.presentation.entertrip.createtrip.choosedate.EnterTripActivity.Companion.TITLE
+import com.going.presentation.entertrip.invitetrip.finish.InviteFinishActivity.Companion.DATE_FORMAT
+import com.going.presentation.entertrip.invitetrip.finish.InviteFinishActivity.Companion.D_DAY_FORMAT
+import com.going.presentation.entertrip.invitetrip.finish.InviteFinishActivity.Companion.TRIP_FORMAT
 import com.going.ui.base.BaseActivity
 import com.going.ui.extension.setOnSingleClickListener
 import com.kakao.sdk.auth.Constants.CODE
