+++ presentation/src/main/res/layout/activity_create_trip.xml
@@ -11,14 +11,14 @@
 
         <variable
             name="viewModel"
-            type="com.going.presentation.entertrip.starttrip.invitetrip.CreateTripViewModel" />
+            type="com.going.presentation.entertrip.invitetrip.invitecode.CreateTripViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white_000"
-        tools:context=".entertrip.starttrip.invitetrip.CreateTripActivity">
+        tools:context=".entertrip.invitetrip.invitecode.CreateTripActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/tb_create_trip"
