+++ presentation/src/main/res/layout/activity_enter_preference.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white_000"
-        tools:context=".entertrip.starttrip.createtrip.EnterPreferenceActivity">
+        tools:context=".entertrip.createtrip.preference.EnterPreferenceActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/toolbar_preference_tag"
