+++ presentation/src/main/res/layout/activity_enter_trip.xml
@@ -11,14 +11,14 @@
 
         <variable
             name="viewModel"
-            type="com.going.presentation.entertrip.starttrip.createtrip.EnterTripViewModel" />
+            type="com.going.presentation.entertrip.createtrip.choosedate.EnterTripViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white_000"
-        tools:context=".entertrip.starttrip.invitetrip.CreateTripActivity">
+        tools:context=".entertrip.invitetrip.invitecode.CreateTripActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/layout_enter_trip"
