+++ presentation/src/main/res/layout/activity_finish_preference.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white_000"
-        tools:context=".entertrip.starttrip.invitetrip.FinishPreferenceActivity">
+        tools:context=".entertrip.invitetrip.preference.FinishPreferenceActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/toolbar_preference_tag"
