+++ presentation/src/main/res/layout/activity_finish_trip.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/gray_50"
-        tools:context=".entertrip.starttrip.createtrip.FinishTripActivity">
+        tools:context=".entertrip.createtrip.finish.FinishTripActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/tb_finish_trip"
