+++ presentation/src/main/res/layout/activity_invite_finish.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/gray_50"
-        tools:context=".entertrip.starttrip.invitetrip.InviteFinishActivity">
+        tools:context=".entertrip.invitetrip.finish.InviteFinishActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/layout_invite_finish_trip"
